+++ src/lib/libheap/heap_msize.c
@@ -1,7 +1,7 @@
 
 #include "_heap.h"
 
-int heap_msize(void* memory)
+size_t heap_msize(void* memory)
 {
     return memory && ((int*)memory)[-1] > sizeof(int) ? ((int*)memory)[-1] - sizeof(int) : 0;
 }
